[WIP] Improve parse error
Attempts to address https://github.com/leanprover/SHerLOC/issues/19
Currently reports the column of the closing quote of the operator.
Not amazing, but also not clear how to get the starting column.
Perhaps it's embedded in the hd of `state.trace` but I haven't
grokked the parser enough yet to know.

Definitely open to feedback on this one!